Added possibility to display menu for D1 Robot.
Whole solution was reworked for better readability.
It is now better represented in real world.
In file `Constants.h` are placed config variables, which define values for display.
Pressing buttons are handled, except for select button.
`Menu` class were introduced, which handles logic of the menu.
It consists of `Categories` class and `Line`.
`D1Robot` class handles hardware shield.
It consists of `LCD` and `AnalogButton` classes, which are handling display and interpretation of pressed button.
DS18B20 class is desired for temperature measurement sensor.